this is my 8th commit

+++ program7.c
@@ -4,7 +4,7 @@ float r;
 printf("enter r");
 scanf("%f",&r);
 
-int area=3.14*r*r;
+float area=3.14*r*r;
 printf("%f", area);
